refactor(router): code refactor
